+++ Assembly-CSharp/RimWorld/BiomeWorker_TemperateForest.cs
@@ -1,5 +1,4 @@
 using RimWorld.Planet;
-using System;
 
 namespace RimWorld
 {
@@ -11,15 +10,15 @@ namespace RimWorld
 			{
 				return -100f;
 			}
-			if (tile.temperature < -10f)
+			if (tile.temperature < -10.0)
 			{
 				return 0f;
 			}
-			if (tile.rainfall < 600f)
+			if (tile.rainfall < 600.0)
 			{
 				return 0f;
 			}
-			return 15f + (tile.temperature - 7f) + (tile.rainfall - 600f) / 180f;
+			return (float)(15.0 + (tile.temperature - 7.0) + (tile.rainfall - 600.0) / 180.0);
 		}
 	}
 }
