+++ Assembly-CSharp/RimWorld/RoomRoleWorker_PrisonBarracks.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using Verse;
 
@@ -35,7 +34,7 @@ namespace RimWorld
 			{
 				return 0f;
 			}
-			return (float)num2 * 100100f + (float)num * 50001f;
+			return (float)((float)num2 * 100100.0 + (float)num * 50001.0);
 		}
 	}
 }
