+++ Assembly-CSharp/Verse/Altitudes.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 
 namespace Verse
@@ -21,13 +20,13 @@ namespace Verse
 			Altitudes.AltIncVect = new Vector3(0f, 0.046875f, 0f);
 			for (int i = 0; i < 31; i++)
 			{
-				Altitudes.Alts[i] = (float)i * 0.46875f;
+				Altitudes.Alts[i] = (float)((float)i * 0.46875);
 			}
 		}
 
 		public static float AltitudeFor(AltitudeLayer alt)
 		{
-			return Altitudes.Alts[(int)alt];
+			return Altitudes.Alts[(uint)alt];
 		}
 	}
 }
